Added CSS and fixed a bug.
Added CSS as to make the game more pleasing to the eye. Also last commit found a bug that sometimes both buttons for Start Game and New Card would stop working. Today figured out that whenever Blackjack is won it wouldn't run New Game as it's set to if hasBlackjack === true it won't draw cards. So had to add a line where it turns hasBlackjack to false during Start Game function.